refactor(frontend/hooks): rename camelCase into kebab-case
